ARM64: dts/qcom: SDM660: Implement RHEL's Low Latency kernel cmdline
mce=ignore_ce - ignores corrected errors and associated scans that can cause
periodic latency spikes.

nosoftlockup - disables logging of backtraces when a process executes on a CPU for
longer than the softlockup threshold (default 120 seconds). Typical low-latency
programming and tuning techniques might involve spinning on a core or modifying
scheduler priorities/policies, which can lead to a task reaching this threshold. If a task
has not relinquished the CPU for 120 seconds, the kernel prints a backtrace for
diagnostic purposes. Adding nosoftlockup to the cmdline disables the printing of this
backtrace (the printk itself can cause latency spikes), and does not in itself reduce
latency. Tasks that spin on a CPU must occasionally yield (especially if they are
SCHED_FIFO priority), or important per-cpu kernel threads may never execute,
potentially leading to unexpected behavior such as very large latency spikes or
interruptions in network traffic.

Documentation: https://access.redhat.com/sites/default/files/attachments/201501-perf-brief-low-latency-tuning-rhel7-v1.1.pdf
Change-Id: Ieaedb58d018ddb85ec46e8fe69729596993aabf1

My edits (@ Tashar02):
As per the in-tree Documentation[1] , we are supposed to disable softlockup with a 0. Disable it properly.

[1] https://github.com/Atom-X-Devs/android_kernel_xiaomi_scarlet/blob/1385eac38ea671b1b1c78b7c1ef1dc92f05d2c32/Documentation/kernel-per-CPU-kthreads.txt#L348

Change-Id: Iba97cf642498240abb33138c0bd4b4e59b32f5e1
Signed-off-by: RyuujiX <saputradenny712@gmail.com>