usb: gadget: f_fs: Avoid use-after-free when disconnect during ongoing IO
There is possibility of race condition between ffs_epfile_io and
ffs_func_unbind when we disconnect a device during ongoing IO.

This could lead to a use-after-free scenario as the ffs_ep
structure would get freed in ffs_func_unbind and trying to access
ep->status in the IPC log could lead to invalid address access.
Prevent this by shifting the IPC log after appropriate checks so
that it can still dump relevant information and not lead to any
use-after-free scenario.

Change-Id: I43af9e78d7246933510a263cbf747286901b3bd6
Signed-off-by: Harsh Agarwal <quic_harshq@quicinc.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: rsuntk <rissu.ntk@gmail.com>