power: supply: qpnp-fg-gen3: fix soc not keeping full after charging overnight
The issue happened when the temperature was cool after charge_done.
(chip->health was not POWER_SUPPLY_HEALTH_GOOD). In this condition,
chip->charge_full was not set, and then battery level dropped.

Set chip->charge_full to true if chip->health is
POWER_SUPPLY_HEALTH_GOOD or POWER_SUPPLY_HEALTH_COOL
or POWER_SUPPLY_HEALTH_WARM to fix it.

Bug: 120606963
Change-Id: Ia762e08826487e626dff775f074205771b908658
Signed-off-by: Jack Wu <wjack@google.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: rssu. <rissu.ntk@gmail.com>